ZNTA-942 make Zanata server an OAuth 2.0 AS and RS
Zanata server now can act as a AS(Authorization Server) and itself being
a RS (Resource Server). It uses authorization code, access token and
refresh token.

This implementation is still a WIP because client id and refresh token
are not persisted to the database. Access token authorization is
temporarily enabled for just one REST resource.